TwoHundredResponder:   - implement TwoHundredResponder and integrate into RouteResponder,   - Add EmptyResourceData null object.   - adjust formatting in DirectoryLinksFormatter.
