+++ Sources/Responders/RouteResponder.swift
@@ -10,7 +10,7 @@ public class RouteResponder: Responder {
   let data: ResourceData
   var logs: [String] = []
 
-  public init(routes: [String: Route], data: ResourceData = ResourceData([:])) {
+  public init(routes: [String: Route], data: ResourceData = EmptyResourceData()) {
     self.routes = routes
     self.data = data
   }
@@ -20,9 +20,13 @@ public class RouteResponder: Responder {
       return HTTPResponse(status: FourHundred.NotFound)
     }
 
-    appendToLogs(request)
+    updateLogsIfValid(request)
 
-    if let clientError = FourHundredResponder(route: route).response(to: request) {
+    let clientErrorResponder = FourHundredResponder(route: route)
+    let redirectResponder = ThreeHundredResponder(redirectPath: route.redirectPath)
+    let successResponder = TwoHundredResponder(route: route, data: data, logs: logs)
+
+    if let clientError = clientErrorResponder.response(to: request) {
       return clientError
     }
 
@@ -30,71 +34,17 @@ public class RouteResponder: Responder {
       return customResponse
     }
 
-    if let redirect = ThreeHundredResponder(route: route).response(to: request) {
+    if let redirect = redirectResponder.response(to: request) {
       return redirect
     }
 
-    return responseByVerb(request: request, route: route)
+    return successResponder.response(to: request)
   }
 
-  private func responseByVerb(request: HTTPRequest, route: Route) -> HTTPResponse {
-    let validResponse = HTTPResponse(status: TwoHundred.Ok)
-    switch request.verb {
-
-    case let verb where verb == .Get:
-      let formatters = getFormatters(request: request, route: route)
-
-      return isAnImage(request.path) ?
-        formatters.first!.addToResponse(validResponse) :
-        formatters.reduce(validResponse) { $1.addToResponse($0) }
-
-    case let verb where verb == .Options:
-      let allowedMethods = route
-                            .allowedMethods
-                            .map { $0.rawValue.uppercased() }
-                            .joined(separator: ",")
-
-      return HTTPResponse(status: TwoHundred.Ok, headers: ["Allow": allowedMethods])
-
-    case let verb where verb == .Post:
-      data.update(request.path, withVal: request.body)
-      return validResponse
-
-    case let verb where verb == .Put:
-      let resource = data.get(request.path)
-      data.update(request.path, withVal: request.body)
-
-      return resource == nil ?
-        HTTPResponse(status: TwoHundred.Created) :
-        validResponse
-
-    case let verb where verb == .Patch:
-      data.update(request.path, withVal: request.body)
-      return HTTPResponse(status: TwoHundred.NoContent)
-
-    case let verb where verb == .Delete:
-      data.remove(at: request.path)
-      return validResponse
-
-    default:
-      return validResponse
-    }
-  }
-
-  private func appendToLogs(_ request: HTTPRequest) {
+  private func updateLogsIfValid(_ request: HTTPRequest) {
     if let verb = request.verb {
       logs.append("\(verb.rawValue.uppercased()) \(request.path) HTTP/1.1")
     }
-  }
-
-  private func getFormatters(request: HTTPRequest, route: Route) -> [ResponseFormatter] {
-    return [
-      ContentFormatter(for: request.path, data: data),
-      route.cookiePrefix.map { CookieFormatter(for: request, prefix: $0) } ?? ParamsFormatter(for: request.params),
-      LogsFormatter(logs: route.includeLogs ? logs : nil),
-      DirectoryLinksFormatter(files: route.includeDirectoryLinks ? data.fileNames : nil),
-      PartialFormatter(for: request.headers["range"])
-    ]
   }
 
 }
