+++ Sources/Responses/String+BytesRepresentable.swift
@@ -11,14 +11,14 @@ public extension String {
     return self.toBytes.toData
   }
 
-  init(response: HTTPResponse) {
+  init?(response: HTTPResponse) {
     let joinedHeaders = response.headers
                                 .map { $0 + response.headerDivide + $1 }
                                 .joined(separator: response.crlf)
 
     let statusLine = "\(response.transferProtocol) \(response.statusCode + response.crlf)"
 
-    self.init(statusLine + joinedHeaders)!
+    self.init(statusLine + joinedHeaders)
   }
 
 }
