+++ Sources/Util/ResourceData.swift
@@ -31,3 +31,9 @@ public class ResourceData {
   }
 
 }
+
+public class EmptyResourceData: ResourceData {
+  public init() {
+    super.init([:])
+  }
+}
